Separate tests in small files, making it easier to edit/add new ones/not conflict on different merge requests
