+++ model_utils/tests/helpers.py
@@ -1,0 +1,5 @@
+
+try:
+    from unittest import skipUnless
+except ImportError: # Python 2.6
+    from django.utils.unittest import skipUnless
